+++ CHANGELOG.rst
@@ -1,7 +1,15 @@
 Changelog
 *********
 
+`Unreleased`_
+-------------
+
+- use simpler commands and allow to set command prefix using "command_prefix" setting in the plugin scope.
+
 1.0.0
 -----
 
 - initial release
+
+
+.. _Unreleased: https://github.com/simplebot-org/simplebot/compare/v1.0.0...HEAD